c-stdaux: make sure c_assert() prints useful information
We want to keep relying on the standard `assert()` macro to print debug
information, but at the same time we want to always evaluate the
argument. Fix the c_assert() symbol to be a macro again so the line
information in case of error is still the line information from the
original source code line.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>