api: add c_memset()
The memset(3) function causes UB if its area pointer is NULL, even if
the area is 0-bytes in length. This is very unfortunate and requires
unnecessary guards in most callers. We really want to be able to call
memset(3) with NULL pointers on empty areas to avoid needless branching
and complexity.

Provide c_memset() which is exactly like memset(3) for non-NULL areas,
but a no-op for empty areas.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>