c-stdaux.h: don't include <stdatomic.h>
c-stdaux is a C11 library, so it's reasonably to require C11 features.
However, RHEL7 comes with gcc 4.8, which has only partial C11 support.
In many aspects it works well enough, except that <stdatomic.h> is missing
([1], [2]).

The includes which "c-stdaux.h" provides to users are part of its API.
It seems the features that <stdatomic.h> provides are not actually used
currently, so drop the include.

gcc-4.8 only has partial C11 support. So even with this change, it's not
that all of c-stdaux works. For example, the unit tests don't compile.
However, there is a commonly used subset of functionality that does in
fact work. To allow users (like NetworkManager) that still build on
RHEL-7, drop <stdatomic.h>.

[1] https://stackoverflow.com/questions/20326604/stdatomic-h-in-gcc-4-8
[2] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=58016
[3] https://gcc.gnu.org/gcc-4.9/changes.html