generic: handle compile time expression in _c_boolean_expr_(),_c_likely_()/_c_unlikely_()
Since recently we use _c_likely_() and _c_boolean_expr_() inside c_assert().
As _c_boolean_expr_() is merely an statement expression, the compiler
cannot recognize constants.

This is for example a problem with c_assert(0) to indicate unreachable
code. With the statement expression, the compiler no longer sees that
the code always fails an assertion (as long as NDEBUG is undefined).
That has consequences for code that follows the statement, which
we know is unreachable. It may lead to compiler warnings like
"-Wsometimes-uninitialized".

Fix that, by evaluating constant expressions explicitly. With this
change, _c_boolean_expr_() can also be used in places where a constant
is required, like `char buffer[_c_boolean_expr_(true) ? 5 : 10];`. Of
course, this is not a useful thing to do directly, but you might use
a macro that ends up effectively calling _c_boolean_expr_() in a context
where a constant is required (like c_assert() using _c_likely_() and
_c_boolean_expr_()).

There is also a unit test that would cause a "-Wsometimes-uninitialized"
warning with clang-15.0.4-1.fc37.

Note that __builtin_choose_expr() is only used with GCC greater than 4.
That's because __builtin_constant_p() inside __builtin_choose_expr()
does not properly work with gcc 4.8 ([2]). It's not too bad. First of
all, c-stdaux is a C11 library, while gcc 4 has only partial C11 support
to being with (it still works in large parts though). Also, the entire
point of _c_boolean_expr_() is to preserve the "-Wparentheses" warning
for a macro argument (while being function-like, evaluating arguments once,
use a __COUNTER__ name, and propagating the constness with
__builtin_choose_expr). It's almost the same as "(!!(_x))" were it not
for "-Wparentheses". We can afford to miss that on non-GCC or on GCC
not newer than 4.

[1] https://github.com/c-util/c-stdaux/pull/11#issuecomment-1331233011
[2] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=19449

Signed-off-by: Thomas Haller <thaller@redhat.com>
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>