c-stdaux: add c_assume_aligned()
New addition to the generic module: c_assume_aligned()

This maps 1-to-1 to the GCC builtin __builtin_assume_aligned(). Refer to
its documentation for details.

This function hints to the compiler that the target of a pointer is
aligned according to the specified alignment. This allows compilers to
use aligned-loads rather than unaligned-loads, as well as other
optimizations.

Since this is an optimization-only, it is mapped to a no-op on platforms
that do not support this builtin.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>