c-stdaux: add c_load*() helpers
Add new c_load*() helper functions, which perform a basic memory load
operation on the specified address+offset. This is effectively identical
to a cast to `uintX_t*` and then dereferencing it. However, the helpers
avoid strict-aliasing-restrictions as well as alignment restrictions.
Hence, they do not violate the rules of the C-standard, unlike the
direct cast.

The functions are designed to allow compiler-optimizations to the
respective multi-byte loads. The ASM output was verified to produce the
expected load-operations.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>